create search custom hook
