+++ src/components/Search/SearchModal.tsx
@@ -1,8 +1,8 @@
-import { env } from "@env/client.mjs";
 import { Dialog } from "@headlessui/react";
 import { MagnifyingGlassIcon, XMarkIcon } from "@heroicons/react/24/outline";
+import useSearch from "@hooks/search";
 import { api, type RouterOutputs } from "@utils/api";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import SearchBookResult from "./SearchBookResult";
 import SearchUserResult from "./SearchUserResult";
 
@@ -29,8 +29,7 @@ const categoryButtonClassName = (
 };
 
 const SearchModal = ({ onCloseDialog, openDialog }: props) => {
-  const [enableSearch, setEnableSearch] = useState(false);
-  const [searchTerm, setSearchTerm] = useState("");
+  const { searchTerm, enableSearch, searchTermChangeHandler } = useSearch();
   const [selectedCategory, setSelectedCategory] =
     useState<SearchCategory>("Users");
   const { data: users } = api.search.searchUsers.useQuery(
@@ -54,15 +53,6 @@ const SearchModal = ({ onCloseDialog, openDialog }: props) => {
       getNextPageParam: (lastPage) => lastPage.nextCursor,
     }
   );
-
-  useEffect(() => {
-    setEnableSearch(false);
-    const delayDebounceFn = setTimeout(() => {
-      setEnableSearch(true);
-    }, env.NEXT_PUBLIC_BOUNCE_DELAY_MILLISECONDS);
-
-    return () => clearTimeout(delayDebounceFn);
-  }, [searchTerm]);
 
   // TODO: handle loading, empty and error state
   const searchResults = (category: SearchCategory) => {
@@ -100,7 +90,7 @@ const SearchModal = ({ onCloseDialog, openDialog }: props) => {
               id="search"
               type="text"
               value={searchTerm}
-              onChange={(event) => setSearchTerm(event.target.value)}
+              onChange={searchTermChangeHandler}
               placeholder="Enter pen name, book title, chapter title..."
             />
           </div>
